+++ packages/common/src/not.ts
@@ -1,2 +1,6 @@
-/** Negates value. */
-export const not = (value: boolean): boolean => !value;
+import { pipe } from '@fluss/core';
+
+/** Negates result of a predicate function. */
+export const not = <T>(
+  predicate: (param: T) => boolean,
+): ((param: T) => boolean) => pipe(predicate, (value: boolean) => !value);