+++ packages/common/test/not.spec.ts
@@ -2,10 +2,10 @@ import { not } from '../src';
 
 describe('not', () => {
   it('should return false if true was received', () => {
-    expect(not(true)).toBe(false);
+    expect(not((_param: unknown) => true)(8)).toBe(false);
   });
 
   it('should return true if false was received', () => {
-    expect(not(false)).toBe(true);
+    expect(not((_param: unknown) => false)(7)).toBe(true);
   });
 });