+++ packages/eleventy-plugin-pwa-icons/src/build_manifest_link.ts
@@ -1,10 +1,10 @@
-import { toRootUrl } from '@eleventy-packages/common';
+import { withLeadingSlash } from '@eleventy-packages/common';
 
 /**
  * Builds link tag of manifest file to be
  * inserted into HTML.
  */
 export const buildManifestLinkTag = (url: string) =>
-  /* html */ `<link rel="manifest" href="${toRootUrl(
+  /* html */ `<link rel="manifest" href="${withLeadingSlash(
     url,
   )}" crossorigin="use-credentials" />`;